Дообавил CallbackQuery и JSONLogger

+++ utils/TGBotApi/User/InterfaceUser.hpp
@@ -8,7 +8,7 @@ using Utils::Types::const_string;
 using Utils::Types::optional_const_string;
 
 struct InterfaceUser {
-    virtual int get_id() const noexcept = 0;
+    virtual long long get_id() const noexcept = 0;
     virtual const_string get_name() const noexcept = 0;
     virtual optional_const_string get_username() const noexcept = 0;
